vlmui/models/CheXagent/requirements.txt
torch==2.6.0
pillow==11.1.0
transformers==4.40.0
matplotlib==3.10.0
albumentations==2.0.4
opencv-python==4.11.0.86
einops==0.8.1
torchvision==0.21.0
pyarrow==19.0.0
accelerate==1.3.0
sentencepiece==0.2.0
protobuf==5.29.3